[Dnssd] Sort IPs by scores for commissionable node discovery
